libreria aadecode
Se extrae el método de openload y videowood a una libreria para hacerlo
más flexible a los cambios.
