+++ python/main-classic/servers/openload.py
@@ -8,7 +8,6 @@
 import re
 from core import scrapertools
 from core import logger
-from core import config
 
            
 headers = {'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.8; rv:22.0) Gecko/20100101 Firefox/22.0',
@@ -37,10 +36,12 @@ def get_video_url(page_url, premium=False, user="", password="", video_password=
         url = page_url.replace("/embed/","/f/")
         data = scrapertools.downloadpageWithoutCookies(url)
         text_encode = scrapertools.get_match(data,"Click to start Download.*?<script[^>]+>(.*?)</script")
-        text_decode = decode(text_encode)
+
     else:
         text_encode = scrapertools.get_match(data,"<video[^<]+<script[^>]+>(.*?)</script>")
-        text_decode = decode(text_encode)
+
+    from aadecode import decode as aadecode
+    text_decode = aadecode(text_encode)
 
     subtitle = scrapertools.find_single_match(data, '<track kind="captions" src="([^"]+)" srclang="es"')
     #Header para la descarga
@@ -84,57 +85,3 @@ def find_videos(text):
             logger.info("  url duplicada=" + url)
 
     return devuelve
-
-def decode(text):
-    text = re.sub(r"\s+", "", text)
-    data = text.split("+(ﾟДﾟ)[ﾟoﾟ]")[1]
-    chars = data.split("+(ﾟДﾟ)[ﾟεﾟ]+")[1:]
-
-
-    txt = ""
-    for char in chars:
-        char = char \
-            .replace("(oﾟｰﾟo)","u") \
-            .replace("c", "0") \
-            .replace("(ﾟДﾟ)['0']", "c") \
-            .replace("ﾟΘﾟ", "1") \
-            .replace("!+[]", "1") \
-            .replace("-~", "1+") \
-            .replace("o", "3") \
-            .replace("_", "3") \
-            .replace("ﾟｰﾟ", "4") \
-            .replace("(+", "(")
-        char = re.sub(r'\((\d)\)', r'\1', char)
-        c = ""; subchar = ""
-        for v in char:
-            c+= v
-            try: x = c; subchar+= str(eval(x)); c = ""
-            except: pass
-        if subchar != '': txt+= subchar + "|"
-    txt = txt[:-1].replace('+','')
-
-    txt_result = "".join([ chr(int(n, 8)) for n in txt.split('|') ])
-    sum_base = ""
-    m3 = False
-    if ".toString(" in txt_result:
-        if "+(" in  txt_result:
-            m3 = True
-            sum_base = "+"+scrapertools.find_single_match(txt_result,".toString...(\d+).")
-            txt_pre_temp = scrapertools.find_multiple_matches(txt_result,"..(\d),(\d+).")
-            txt_temp = [ (n, b) for b ,n in txt_pre_temp ]
-        else:
-            txt_temp = scrapertools.find_multiple_matches(txt_result, '(\d+)\.0.\w+.([^\)]+).')
-        for numero, base in txt_temp:
-            code = toString( int(numero), eval(base+sum_base) )
-            if m3:
-                txt_result = re.sub( r'"|\+', '', txt_result.replace("("+base+","+numero+")", code) )
-            else:
-                txt_result = re.sub( r"'|\+", '', txt_result.replace(numero+".0.toString("+base+")", code) )
-    return txt_result
-
-def toString(number,base):
-   string = "0123456789abcdefghijklmnopqrstuvwxyz"
-   if number < base:
-      return string[number]
-   else:
-      return toString(number//base,base) + string[number%base]
